requirements-ci.txt
python-dotenv
requests
pandas
pyarrow
hopsworks
nbconvert
ipykernel